+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Show total fiat amount in Bitrefill confirmation step
 - Show contract address in Ethereum token account info
 - Harmonize iOS and Android back swipe/button navigation to only work in expected situations
+- Fix coin control reused-address labels
 
 ## v4.50.1
 - Fix a bug that would delay showing watch-only accounts.